New logo! Optimizations galore! Refactoring! Too much to mention!

+++ soundboard/src/main/java/us/huseli/soundboard2/data/repositories/CategoryRepository.kt
@@ -1,10 +1,9 @@
 package us.huseli.soundboard2.data.repositories
 
-import android.content.Context
-import androidx.core.net.toUri
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.flow.*
-import us.huseli.soundboard2.Constants
+import androidx.annotation.ColorInt
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.map
 import us.huseli.soundboard2.data.dao.CategoryDao
 import us.huseli.soundboard2.data.dao.SoundDao
 import us.huseli.soundboard2.data.entities.Category
@@ -18,57 +17,42 @@ import javax.inject.Singleton
 class CategoryRepository @Inject constructor(
     private val categoryDao: CategoryDao,
     private val soundDao: SoundDao,
-    colorHelper: ColorHelper,
-    @ApplicationContext private val context: Context
+    private val colorHelper: ColorHelper
 ) : LoggingObject {
     val categories: Flow<List<Category>> = categoryDao.flowList()
     val categoryIds: Flow<List<Int>> = categoryDao.flowListIds()
-    val randomColor: Flow<Int> = categoryDao.flowListUsedColors().map { colorHelper.getRandomColor(exclude = it) }
-
     val firstCategory: Flow<Category> = categories.map {
         if (it.isEmpty()) createDefault()
         it.firstOrNull()
     }.filterNotNull()
 
-    fun get(categoryId: Int): Flow<Category?> = categoryDao.flowGet(categoryId)
-    fun getSoundCount(categoryId: Int): Flow<Int> = categoryDao.flowGetSoundCount(categoryId)
+    fun flowGet(categoryId: Int): Flow<Category?> = categoryDao.flowGet(categoryId)
     fun isFirstCategory(categoryId: Int): Flow<Boolean> = categoryIds.map { it.firstOrNull() == categoryId }
     fun isLastCategory(categoryId: Int) = categoryIds.map { it.lastOrNull() == categoryId }
+
+    suspend fun get(categoryId: Int): Category = categoryDao.get(categoryId)
+    suspend fun getRandomColor(@ColorInt vararg exclude: Int): Int =
+        colorHelper.getRandomColor(exclude = categoryDao.listUsedColors() + exclude.asList())
+
+    suspend fun getSoundCount(categoryId: Int): Int = categoryDao.getSoundCount(categoryId)
+    suspend fun list(): List<Category> = categoryDao.list()
     suspend fun toggleCollapsed(categoryId: Int) = categoryDao.toggleCollapsed(categoryId)
-    suspend fun create(name: CharSequence, backgroundColor: Int, soundSorting: SoundSorting) =
+    suspend fun create(name: CharSequence, @ColorInt backgroundColor: Int, soundSorting: SoundSorting) =
         categoryDao.create(name.toString(), backgroundColor, categoryDao.getNextPosition(), soundSorting)
 
     suspend fun delete(category: Category, moveSoundsTo: Int?) {
-        val sounds = soundDao.listByCategoryId(category.id)
-
-        if (moveSoundsTo == null) {
-            val duplicateUris = soundDao.listByChecksums(sounds.map { it.checksum })
-                .filter { it.categoryId != category.id }
-                .map { it.uri }
-                .toSet()
-            val urisToDelete = sounds.map { it.uri }.subtract(duplicateUris)
-
-            context.getDir(Constants.SOUND_DIRNAME, Context.MODE_PRIVATE)
-                ?.listFiles()
-                ?.forEach { if (urisToDelete.contains(it.toUri())) it.delete() }
-            soundDao.deleteByCategoryId(category.id)
-        }
-        else {
-            val newSounds = sounds.map { sound ->
-                sound.clone(categoryId = moveSoundsTo)
-            }
-            soundDao.update(newSounds)
-        }
-        categoryDao.delete(category)
+        if (moveSoundsTo == null) soundDao.deleteByCategoryId(category.id)
+        else soundDao.update(soundDao.listByCategoryId(category.id).map { sound ->
+            sound.clone(categoryId = moveSoundsTo)
+        })
+        categoryDao.delete(listOf(category))
     }
-
-    // suspend fun update(category: Category) = categoryDao.update(category)
 
     suspend fun update(vararg categories: Category) = categoryDao.update(categories.asList())
 
     suspend fun createDefault() = create(
         "Dëfäult",
-        randomColor.first(),
+        getRandomColor(),
         SoundSorting(SoundSorting.Parameter.NAME, SoundSorting.Order.ASCENDING)
     )
 }
